+++ src/reportlab/__init__.py
@@ -1,9 +1,9 @@
 #Copyright ReportLab Europe Ltd. 2000-2023
 #see license.txt for license details
 __doc__="""The Reportlab PDF generation library."""
-Version = "4.4.4"
+Version = "4.4.5"
 __version__=Version
-__date__='20250918'
+__date__='20251117'
 
 import sys, os