+++ CHANGES.md
@@ -11,11 +11,12 @@ E.g. to retrieve the changes made between release 3.4 and release 3.5, type::
 The contributors lists are in no order and apologies to those accidentally not
 mentioned. If we missed you, please let us know!
 
-CHANGES  4.3.5   ??/??/2025
+CHANGES  4.3.5   17/11/2025
 ---------------------------
     * remove random monkey patches in randomtext
     * add and use testutils.invariantSeed in tests
     * fix (maybe partially) Table row splitting of ListFlowable
+    * apply patch for in row splitting bug reported by Christian Zwicknagl via Yoshua Wakeham
 
 CHANGES  4.3.4   18/09/2025
 ---------------------------
@@ -1107,6 +1108,8 @@ RELEASE 2.4  at 18:00 GMT  20/Jan/2010
    * Michael Egorov <michwill@gmail.com>
    * Mike Folwell <mjf@pearson.co.uk>
    * Robert Alsina
+   * Christian Zwicknagl
+   * Yoshua Wakeham
    * and more ...